Create shipment reducer, move cargo reducer into its own folder (separate from quote)
