+++ imports/state/reducers/shipment/shipmentReducer.js
@@ -1,0 +1,5 @@
+import { cargo } from '../cargo/cargoReducers';
+
+export const shipment = (state = {}, action) => ({
+  cargo: cargo(state.cargo || {}, action),
+});
